third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-1st-stop-expected.txt

1	2	3
4	5	6
7	8	9
PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
